requirements.txt
streamlit
pillow
requests
python-dotenv
pandas
google-generativeai
openpyxl